chore(refacto): Split tests in files + Combine flags
