+++ generate/generate.go
@@ -6,43 +6,46 @@ import (
 	"strings"
 )
 
+// WithSpecialChars adds special characters to each word in the wordlist
 func WithSpecialChars(wordlist []string) []string {
+	var result []string
 	specialChars := []string{"!", "@", "#", "$", "%", "+", "?", "=", "*"}
 	for _, word := range wordlist {
 		for _, char := range specialChars {
-			wordlist = append(wordlist, word+char)
+			result = append(result, word+char)
 		}
 	}
-
-	return wordlist
+	return result
 }
 
+// WithNumbers appends numbers to each word in the wordlist up to a specified range
 func WithNumbers(wordlist []string, specificRange int) []string {
-	Numbers := make([]int, specificRange+1)
+	var result []string
 	for _, word := range wordlist {
-		for i := range Numbers {
-			newPassword := word + strconv.Itoa(i)
-
-			wordlist = append(wordlist, newPassword)
+		for i := 0; i <= specificRange; i++ {
+			result = append(result, word+strconv.Itoa(i))
 		}
 	}
-	return wordlist
+	return result
 }
 
+// WithMixedCase generates case variations for each word in the wordlist
 func WithMixedCase(wordlist []string) []string {
+	var result []string
 	for _, word := range wordlist {
-		caseVariations := generateCaseVariations(word)
-		wordlist = append(wordlist, caseVariations...)
+		result = append(result, generateCaseVariations(word)...)
 	}
-	return wordlist
+	return result
 }
 
+// generateCaseVariations generates all possible case variations of a word
 func generateCaseVariations(word string) []string {
 	var result []string
 	helper(word, "", 0, &result)
 	return result
 }
 
+// helper is a recursive function that generates case variations of a word
 func helper(word, current string, index int, result *[]string) {
 	if index == len(word) {
 		*result = append(*result, current)
@@ -56,14 +59,15 @@ func helper(word, current string, index int, result *[]string) {
 	}
 }
 
-func WithYearAndSeparators(wordlist []string, year int, separators string, yearRange int) []string {
-	var newWordlist []string
+// WithYear appends the year to each word in the wordlist
+func WithYear(wordlist []string, year int, yearRange int, separators string) []string {
+	var result []string
 	separatorsList := strings.Split(separators, "")
 
 	addYearWithSeparators := func(word string, year int) {
-		newWordlist = append(newWordlist, fmt.Sprintf("%s%d", word, year))
+		result = append(result, fmt.Sprintf("%s%d", word, year))
 		for _, separator := range separatorsList {
-			newWordlist = append(newWordlist, fmt.Sprintf("%s%s%d", word, separator, year))
+			result = append(result, fmt.Sprintf("%s%s%d", word, separator, year))
 		}
 	}
 
@@ -79,21 +83,25 @@ func WithYearAndSeparators(wordlist []string, year int, separators string, yearR
 		}
 	}
 
-	return append(wordlist, newWordlist...)
+	return append(wordlist, result...)
 }
 
+// WithStartCaps capitalizes the first letter of each word in the wordlist
 func WithStartCaps(wordlist []string) []string {
+	var result []string
 	for _, word := range wordlist {
 		if len(word) > 0 {
-			wordlist = append(wordlist, strings.ToUpper(string(word[0]))+word[1:])
+			result = append(result, strings.ToUpper(string(word[0]))+word[1:])
 		} else {
-			wordlist = append(wordlist, word)
+			result = append(result, word)
 		}
 	}
-	return wordlist
+	return result
 }
 
+// WithLeetCode converts certain characters in each word to leet speak
 func WithLeetCode(wordlist []string) []string {
+	var result []string
 	leetMap := map[rune][]string{
 		'a': {"4", "@"}, 'A': {"4", "@"},
 		'e': {"3"}, 'E': {"3"},
@@ -105,19 +113,20 @@ func WithLeetCode(wordlist []string) []string {
 	}
 
 	for _, word := range wordlist {
-		leetVariations := generateLeetVariations(word, leetMap)
-		wordlist = append(wordlist, leetVariations...)
+		result = append(result, generateLeetVariations(word, leetMap)...)
 	}
 
-	return wordlist
+	return result
 }
 
+// generateLeetVariations generates all possible leet speak variations of a word
 func generateLeetVariations(word string, leetMap map[rune][]string) []string {
 	var result []string
 	helperLeet(word, "", 0, &result, leetMap)
 	return result
 }
 
+// helperLeet is a recursive function that generates leet speak variations of a word
 func helperLeet(word, current string, index int, result *[]string, leetMap map[rune][]string) {
 	if index == len(word) {
 		*result = append(*result, current)
@@ -133,7 +142,9 @@ func helperLeet(word, current string, index int, result *[]string, leetMap map[r
 	helperLeet(word, current+string(char), index+1, result, leetMap)
 }
 
+// WithMask applies a custom mask to each word in the wordlist
 func WithMask(wordlist []string, mask string) []string {
+	var result []string
 	specialChars := []rune("!@#$%^&*()_+-=[]{}|;:'\",.<>?/\\")
 	digits := []rune("0123456789")
 	lowercase := []rune("abcdefghijklmnopqrstuvwxyz")
@@ -143,12 +154,13 @@ func WithMask(wordlist []string, mask string) []string {
 	for _, word := range wordlist {
 		var maskedWords []string
 		applyMask(word, mask, "", &maskedWords, specialChars, digits, lowercase, uppercase, allChars)
-		wordlist = append(wordlist, maskedWords...)
+		result = append(result, maskedWords...)
 	}
 
-	return wordlist
+	return result
 }
 
+// applyMask is a recursive function that applies a custom mask to a word
 func applyMask(word, mask, current string, result *[]string, specialChars, digits, lowercase, uppercase, allChars []rune) {
 	if len(mask) == 0 {
 		*result = append(*result, current)
@@ -188,16 +200,20 @@ func applyMask(word, mask, current string, result *[]string, specialChars, digit
 	}
 }
 
+// WithUppercase converts each word in the wordlist to uppercase
 func WithUppercase(wordlist []string) []string {
+	var result []string
 	for _, word := range wordlist {
-		wordlist = append(wordlist, strings.ToUpper(word))
+		result = append(result, strings.ToUpper(word))
 	}
-	return wordlist
+	return result
 }
 
+// WithPostal appends the postal code to each word in the wordlist
 func WithPostal(wordlist []string, includePostal int) []string {
+	var result []string
 	for _, word := range wordlist {
-		wordlist = append(wordlist, word+strconv.Itoa(includePostal))
+		result = append(result, word+strconv.Itoa(includePostal))
 	}
-	return wordlist
+	return result
 }
